Add Token to Number Conversion Extensions
